Rename DiagnosticLogging to DiagnosticLogger
