+++ JavaScriptSDK/DiagnosticLogger.ts
